backend/src/requirements.txt
psycopg2-binary==2.9.9
pgvector==0.2.4
voyageai
google-generativeai==0.3.2
python-dotenv==1.0.1
langchain==0.1.5
langchain-community
langchain-core==0.1.17
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1 
psycopg2